fix rubocop complains
