+++ lib/capybara/pagemap/configuration.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 module Capybara
+  # Pagemap module extend functionality to clases to map capybara nodes
   module Pagemap
     class << self
       attr_writer :configuration
@@ -14,6 +15,7 @@ module Capybara
       yield(configuration)
     end
 
+    # Configuration class encapsulates gobal configurations
     class Configuration
       attr_accessor :enabled
       def initialize
