+++ test/capybara/pagemap/pagemap_test.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+module Capybara
+  class PagemapTest < Minitest::Test
+    def test_that_it_has_a_version_number
+      refute_nil ::Capybara::Pagemap::VERSION
+    end
+  end
+end
